IPRT/rest: build fix. bugref:9167
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@74196 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/testcase/tstRTCRest-1.cpp
@@ -1718,7 +1718,7 @@ static void testArray()
                 else
                     RTTESTI_CHECK_RC(Arr4.prepend(pDirect), VINF_SUCCESS);
             }
-            RTTESTI_CHECK(Arr4.size() == i + 1);
+            RTTESTI_CHECK((ssize_t)Arr4.size() == i + 1);
             RTTESTI_CHECK(Arr4.isEmpty() == false);
         }
